Move local_planner_node main() to its own file, to ease testing local_planner_node
