+++ local_planner/CMakeLists.txt
@@ -138,7 +138,8 @@ add_dependencies(local_planner ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPO
 
 ## Declare a C++ executable
 # add_executable(avoidance_node src/avoidance_node.cpp)
-add_executable(local_planner_node src/nodes/local_planner_node.cpp)
+add_executable(local_planner_node src/nodes/local_planner_node.cpp
+                                  src/nodes/local_planner_node_main.cpp)
 
 ## Add cmake target dependencies of the executable
 ## same as for the library above
